Move environment runner to scripts
